chore: remove windows build warnings

+++ source/nvim/api/buffer.c
@@ -431,7 +431,7 @@ FUNC_API_SINCE(1)
     // the buffer, otherwise line numbers will be invalid.
     if(save_curbuf.br_buf == NULL)
     {
-        mark_adjust((linenr_T)start, (linenr_T)(end - 1), MAXLNUM, extra);
+        mark_adjust((linenr_T)start, (linenr_T)(end - 1), MAXLNUM, (long)extra);
     }
 
     changed_lines((linenr_T)start, 0, (linenr_T)end, (long)extra);
